requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
google-generativeai==0.8.5
google-cloud-firestore==2.19.0
google-cloud-texttospeech==2.21.1
pydantic==2.11.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.25.3
